Add files via upload
This is a simple pong game created using python and the Pygame module in python. The game features a player and the CPU playing against each other where both the CPU and the player are given paddles and play with a ball. Player who fails to block the ball from hitting the wall loses.
This project covers a lot of the aspects of computer game programming, like movement, control, collision detection, scoring, artificial intelligence and so on.